+++ src/Haskellorls/NodeInfo.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RecordWildCards #-}
 
@@ -6,34 +7,45 @@ module Haskellorls.NodeInfo
     nodeInfo,
     nodeInfoStatus,
     nodeInfoPath,
+    nodeInfoContext,
     toFileInfo,
   )
 where
 
+import qualified Control.Exception.Base as Exception
+import qualified Data.Either as Either
 import qualified Data.List as L
+import qualified Data.Text as T
 import qualified Haskellorls.Option as Option
 import qualified Haskellorls.Tree.Type as Tree
 import qualified Haskellorls.Utils as Utils
 import qualified System.FilePath.Posix as Posix
 import qualified System.IO as IO
 import qualified System.Posix.Files as Files
+#ifdef SELINUX
+import qualified System.Linux.SELinux as SELinux
+#endif
 
 data NodeInfo
   = FileInfo
       { getFilePath :: FilePath,
         getFileStatus :: Files.FileStatus,
+        getFileContext :: T.Text,
         getTreeNodePositions :: [Tree.TreeNodePosition]
       }
   | LinkInfo
       { getLinkPath :: FilePath,
         getLinkStatus :: Files.FileStatus,
+        getLinkContext :: T.Text,
         getDestPath :: FilePath,
         getDestStatus :: Files.FileStatus,
+        getDestContext :: T.Text,
         getTreeNodePositions :: [Tree.TreeNodePosition]
       }
   | OrphanedLinkInfo
       { getOrphanedLinkPath :: FilePath,
         getOrphanedLinkStatus :: Files.FileStatus,
+        getOrphanedLinkContext :: T.Text,
         getDestPath :: FilePath,
         getTreeNodePositions :: [Tree.TreeNodePosition]
       }
@@ -41,6 +53,7 @@ data NodeInfo
 nodeInfo :: Option.Option -> FilePath -> FilePath -> IO NodeInfo
 nodeInfo opt dirname basename = do
   status <- Files.getSymbolicLinkStatus path
+  context <- fileContext path
   if Files.isSymbolicLink status
     then do
       linkPath <- Utils.readSymbolicLink path
@@ -53,11 +66,17 @@ nodeInfo opt dirname basename = do
           Right p -> Utils.destFileStatus $ linkDestPath path p
           _ -> pure Nothing
 
+      destContext <- do
+        case linkPath of
+          Right p -> fileContext p
+          _ -> pure defaultContext
+
       return $ case (linkPath, destStatus) of
         (Right p, Nothing) ->
           OrphanedLinkInfo
             { getOrphanedLinkPath = basename,
               getOrphanedLinkStatus = status,
+              getOrphanedLinkContext = T.pack context,
               getDestPath = p,
               getTreeNodePositions = []
             }
@@ -68,20 +87,24 @@ nodeInfo opt dirname basename = do
             FileInfo
               { getFilePath = basename,
                 getFileStatus = s,
+                getFileContext = T.pack context,
                 getTreeNodePositions = []
               }
           | otherwise ->
             LinkInfo
               { getLinkPath = basename,
                 getLinkStatus = status,
+                getLinkContext = T.pack context,
                 getDestPath = p,
                 getDestStatus = s,
+                getDestContext = T.pack destContext,
                 getTreeNodePositions = []
               }
         _ ->
           FileInfo
             { getFilePath = basename,
               getFileStatus = status,
+              getFileContext = T.pack context,
               getTreeNodePositions = []
             }
     else
@@ -89,6 +112,7 @@ nodeInfo opt dirname basename = do
         FileInfo
           { getFilePath = basename,
             getFileStatus = status,
+            getFileContext = T.pack context,
             getTreeNodePositions = []
           }
   where
@@ -101,12 +125,14 @@ toFileInfo = \case
     FileInfo
       { getFilePath = getDestPath,
         getFileStatus = getDestStatus,
+        getFileContext = getDestContext,
         ..
       }
   OrphanedLinkInfo {..} ->
     FileInfo
       { getFilePath = getDestPath,
         getFileStatus = getOrphanedLinkStatus,
+        getFileContext = getOrphanedLinkContext,
         ..
       }
 
@@ -129,3 +155,23 @@ nodeInfoPath = \case
   FileInfo {..} -> getFilePath
   LinkInfo {..} -> getLinkPath
   OrphanedLinkInfo {..} -> getOrphanedLinkPath
+
+nodeInfoContext :: NodeInfo -> T.Text
+nodeInfoContext = \case
+  FileInfo {..} -> getFileContext
+  LinkInfo {..} -> getLinkContext
+  OrphanedLinkInfo {..} -> getOrphanedLinkContext
+
+-- | NOTE: This is not tested on SELinux enabled environment so maybe break.
+fileContext :: FilePath -> IO String
+fileContext path =
+#ifdef SELINUX
+  do
+    context <- Exception.try (SELinux.getFileCon path) :: IO (Either Exception.IOException String)
+    pure $ Either.fromRight defaultContext context
+#else
+  pure defaultContext
+#endif
+
+defaultContext :: String
+defaultContext = "?"